Define a Travis build leveraging GH Pages provider
This builds on the definitions of auto-deployment wihtin CyVerse-UI.

It adds a Travis CI build definition that will install `pandoc`, then
generate an HTML representation of the overall "guides". It makes use
of an existing Deployment Provider, pages [1], offered by Travis CI. On
success, this will be deployed to GitHub Pages by Travis CI.

The "auto-deploy" is performed by a "service" account, @cyverse-deploy.
@cyverse-deploy has *not* be added to the @cyverse org to avoid any leak
of permissions. This add only has "write" permission to
`cyverse/atmosphere-guides`, by **design**.

[0] https://github.com/cyverse/cyverse-ui/pull/40/commits/4f6cd26de021ccd90fbb320be80762a26354377a
[1] https://docs.travis-ci.com/user/deployment/pages/