Scope travis deployment to only occur on cyverse/atmosphere-guides repo
Currently every fork of this guide will fail travis, because it has a built in
encrypted secret GITHUB_TOKEN for atmosphere's deployment agent. The
deployment agent only has access to a handful of repos.

This fix means that forks will not try to build gh-pages. In the future it
would be nice to allow that automatically. But if you have a fork its easy
enough to push up to your own gh-pages manually to test.